Fixed docs problems when they are built in production
Fixed images and fonts that are served from a static folder to be visible. Enabled hugo-stats.json file for the postcss step. Added new shortcodes for setting link, script and generating url that are with source from the static folder. Disabled changelog.  Fixed transfer docs script to transfer md files in correct folders.