Dependabot now merges its opened PRs
Also it will label with 'requires-manual-qa' the PRs that are having major updates and will comment to the PR that it was not approved. Will auto add 'dependencies' label to the PR if not added because is needed when releasing a new component version the dependencies updates to be marked as 'Dependencies' and not 'Features' to the release notes.